+++ source/extensions/isaacsim.core.simulation_manager/python/tests/test_extension.py
@@ -58,3 +58,73 @@ class TestExtension(omni.kit.test.AsyncTestCase):
         await update_stage_async()
         for callback_id in self._callbacks:
             SimulationManager.deregister_callback(callback_id)
+
+    async def test_physics_callbacks(self):
+        await create_new_stage_async()
+        global var
+        var = 1
+
+        def add_one(dt):
+            global var
+            var += 1
+
+        def multiply_two(dt):
+            global var
+            var *= 2
+
+        self._callbacks = []
+        self._callbacks.append(SimulationManager.register_callback(add_one, event=IsaacEvents.PRE_PHYSICS_STEP))
+        self._callbacks.append(
+            SimulationManager.register_callback(add_one, event=IsaacEvents.POST_PHYSICS_STEP, order=2)
+        )
+        self._callbacks.append(
+            SimulationManager.register_callback(multiply_two, event=IsaacEvents.POST_PHYSICS_STEP, order=0)
+        )
+        timeline = omni.timeline.get_timeline_interface()
+        timeline.play()
+        timeline.commit()
+        SimulationManager.step(render=False)
+        SimulationManager.step(render=False)
+        for callback_id in self._callbacks:
+            SimulationManager.deregister_callback(callback_id)
+        self.assertEqual(var, 13)
+
+    async def test_simulation_manager_interface(self):
+        timeline = omni.timeline.get_timeline_interface()
+        await create_new_stage_async()
+
+        self.assertEqual(SimulationManager.get_num_physics_steps(), 0)
+        self.assertEqual(SimulationManager.get_simulation_time(), 0.0)
+        self.assertEqual(SimulationManager.is_simulating(), False)
+        self.assertEqual(SimulationManager.is_paused(), False)
+        self.assertEqual(SimulationManager.get_physics_dt(), 1.0 / 60.0)
+
+        timeline.play()
+        await update_stage_async()
+        self.assertEqual(SimulationManager.is_simulating(), True)
+        self.assertEqual(SimulationManager.is_paused(), False)
+        # The number of steps is offset by 2 because we do two steps of simulation to warm up
+        self.assertEqual(SimulationManager.get_num_physics_steps(), 3)
+        self.assertAlmostEqual(SimulationManager.get_simulation_time(), 1.0 / 60.0 * 3)
+
+        timeline.pause()
+        await update_stage_async()
+        self.assertEqual(SimulationManager.is_simulating(), True)
+        self.assertEqual(SimulationManager.is_paused(), True)
+        self.assertEqual(SimulationManager.get_num_physics_steps(), 3)
+        self.assertAlmostEqual(SimulationManager.get_simulation_time(), 1.0 / 60.0 * 3)
+
+        timeline.play()
+        await update_stage_async()
+        await update_stage_async()
+        self.assertEqual(SimulationManager.is_simulating(), True)
+        self.assertEqual(SimulationManager.is_paused(), False)
+        self.assertEqual(SimulationManager.get_num_physics_steps(), 5)
+        self.assertAlmostEqual(SimulationManager.get_simulation_time(), 1.0 / 60.0 * 5)
+
+        timeline.stop()
+        await update_stage_async()
+        self.assertEqual(SimulationManager.is_simulating(), False)
+        self.assertEqual(SimulationManager.is_paused(), False)
+        self.assertEqual(SimulationManager.get_num_physics_steps(), 0)
+        self.assertEqual(SimulationManager.get_simulation_time(), 0)